ADJST-817 Using search api to find prisoner details instead of prison-api.

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsidentifyremandperiodsapi/relevantremand/controller/RelevantRemandController.kt
@@ -59,7 +59,7 @@ class RelevantRemandController(
     log.info("Request received to calculate relevant remand for $prisonerId")
     val courtDateResults = prisonService.getCourtDateResults(prisonerId)
     val prisonerDetails = prisonService.getOffenderDetail(prisonerId)
-    val sentencesAndOffences = prisonService.getSentencesAndOffences(prisonerDetails.bookingId, true)
+    val sentencesAndOffences = prisonService.getSentencesAndOffences(prisonerDetails.bookingId.toLong(), true)
     return remandCalculationService.calculate(transform(courtDateResults, prisonerDetails, sentencesAndOffences), remandCalculationRequestOptions)
   }
